Add files via upload
Runs in command prompt; must pip install numpy and scipy.io to run in the prompt. Also nests functions so that you only have to run one function main() to take csv files, convert frames to timestamps, and export as a MATLAB array file, which offers inputs for names of files and removes need to label file types.